Motor control new
Change pd algorithym

+++ stm32flash/src/math.c
@@ -8,14 +8,14 @@ int abs(int num){
     return num;
 }
 
-short int twos16Bit(uint16_t num){
+int twos16Bit(uint16_t num){
     if(num >> 15){
         num = ~num;
         num++;
-        return num*(-1);
+        return (int)num*(-1);
     }
     else{
-        return (short int)num;
+        return (int)num;
     }
 }
